+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2019-12-16
+
+### Changed
+
+- `LocaleField::getLocales()` is now in the config file as `nova-locale-field.locales`
+
+### Removed
+
+- `LocaleChildrenFilter` is now deprecated and replaced with just `LocaleFilter` which uses the first locale as the default value
+
 ## [1.1.4] - 2019-11-11
 
 ### Changed
@@ -60,6 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Default locales support
 - Value override support (locales, max locales on index)
 
+[2.0.0]: https://github.com/optimistdigital/nova-locale-field/compare/1.1.4...2.0.0
 [1.1.4]: https://github.com/optimistdigital/nova-locale-field/compare/1.1.3...1.1.4
 [1.1.3]: https://github.com/optimistdigital/nova-locale-field/compare/1.1.2...1.1.3
 [1.1.2]: https://github.com/optimistdigital/nova-locale-field/compare/1.1.1...1.1.2